Use Step, PickleStep and Hook naming convention
To stay consistent with the TestStep, PickleStepTestStep and HookStep
from `cucumber.api` we should use the same naming convention in
`cucumber.runtime`.
