+++ core/src/test/java/cucumber/runner/HookTestStepTest.java
@@ -5,7 +5,6 @@ import cucumber.api.Result;
 import cucumber.api.Scenario;
 import cucumber.api.event.TestStepFinished;
 import cucumber.api.event.TestStepStarted;
-import cucumber.runtime.DefinitionMatch;
 import cucumber.runtime.HookDefinitionMatch;
 import org.junit.Test;
 import org.mockito.InOrder;
