+++ core/src/test/java/cucumber/runtime/RuntimeTest.java
@@ -255,7 +255,7 @@ public class RuntimeTest {
     @Test
     public void should_add_passed_result_to_the_summary_counter() throws Exception {
         ByteArrayOutputStream baos = new ByteArrayOutputStream();
-        StepDefinitionMatch match = mock(StepDefinitionMatch.class);
+        PickleStepDefinitionMatch match = mock(PickleStepDefinitionMatch.class);
 
         Runtime runtime = createRuntimeWithMockedGlue(match, "--monochrome");
         runScenario(runtime, stepCount(1));
@@ -269,7 +269,7 @@ public class RuntimeTest {
     @Test
     public void should_add_pending_result_to_the_summary_counter() throws Throwable {
         ByteArrayOutputStream baos = new ByteArrayOutputStream();
-        StepDefinitionMatch match = createExceptionThrowingMatch(new PendingException());
+        PickleStepDefinitionMatch match = createExceptionThrowingMatch(new PendingException());
 
         Runtime runtime = createRuntimeWithMockedGlue(match, "--monochrome");
         runScenario(runtime, stepCount(1));
@@ -283,7 +283,7 @@ public class RuntimeTest {
     @Test
     public void should_add_failed_result_to_the_summary_counter() throws Throwable {
         ByteArrayOutputStream baos = new ByteArrayOutputStream();
-        StepDefinitionMatch match = createExceptionThrowingMatch(new Exception());
+        PickleStepDefinitionMatch match = createExceptionThrowingMatch(new Exception());
 
         Runtime runtime = createRuntimeWithMockedGlue(match, "--monochrome");
         runScenario(runtime, stepCount(1));
@@ -310,7 +310,7 @@ public class RuntimeTest {
     @Test
     public void should_add_skipped_result_to_the_summary_counter() throws Throwable {
         ByteArrayOutputStream baos = new ByteArrayOutputStream();
-        StepDefinitionMatch match = createExceptionThrowingMatch(new Exception());
+        PickleStepDefinitionMatch match = createExceptionThrowingMatch(new Exception());
 
         Runtime runtime = createRuntimeWithMockedGlue(match, "--monochrome");
         runScenario(runtime, stepCount(2));
@@ -337,7 +337,7 @@ public class RuntimeTest {
     @Test
     public void should_fail_the_scenario_if_before_fails() throws Throwable {
         ByteArrayOutputStream baos = new ByteArrayOutputStream();
-        StepDefinitionMatch match = mock(StepDefinitionMatch.class);
+        PickleStepDefinitionMatch match = mock(PickleStepDefinitionMatch.class);
         HookDefinition hook = createExceptionThrowingHook();
 
         Runtime runtime = createRuntimeWithMockedGlue(match, hook, HookType.Before, "--monochrome");
@@ -352,7 +352,7 @@ public class RuntimeTest {
     @Test
     public void should_fail_the_scenario_if_after_fails() throws Throwable {
         ByteArrayOutputStream baos = new ByteArrayOutputStream();
-        StepDefinitionMatch match = mock(StepDefinitionMatch.class);
+        PickleStepDefinitionMatch match = mock(PickleStepDefinitionMatch.class);
         HookDefinition hook = createExceptionThrowingHook();
 
         Runtime runtime = createRuntimeWithMockedGlue(match, hook, HookType.After, "--monochrome");
@@ -375,7 +375,7 @@ public class RuntimeTest {
         HookDefinition beforeHook = mock(HookDefinition.class);
         when(beforeHook.matches(anyCollectionOf(PickleTag.class))).thenReturn(true);
 
-        Runtime runtime = createRuntimeWithMockedGlue(mock(StepDefinitionMatch.class), beforeHook, HookType.Before);
+        Runtime runtime = createRuntimeWithMockedGlue(mock(PickleStepDefinitionMatch.class), beforeHook, HookType.Before);
         runtime.runFeature(feature);
 
         ArgumentCaptor<Scenario> capturedScenario = ArgumentCaptor.forClass(Scenario.class);
@@ -476,8 +476,8 @@ public class RuntimeTest {
         return formatterSpy.toString();
     }
 
-    private StepDefinitionMatch createExceptionThrowingMatch(Exception exception) throws Throwable {
-        StepDefinitionMatch match = mock(StepDefinitionMatch.class);
+    private PickleStepDefinitionMatch createExceptionThrowingMatch(Exception exception) throws Throwable {
+        PickleStepDefinitionMatch match = mock(PickleStepDefinitionMatch.class);
         doThrow(exception).when(match).runStep(anyString(), (Scenario) any());
         return match;
     }
@@ -521,20 +521,20 @@ public class RuntimeTest {
         return new Runtime(resourceLoader, classLoader, backends, runtimeOptions);
     }
 
-    private Runtime createRuntimeWithMockedGlue(StepDefinitionMatch match, String... runtimeArgs) {
+    private Runtime createRuntimeWithMockedGlue(PickleStepDefinitionMatch match, String... runtimeArgs) {
         return createRuntimeWithMockedGlue(match, false, mock(HookDefinition.class), HookType.After, runtimeArgs);
     }
 
-    private Runtime createRuntimeWithMockedGlue(StepDefinitionMatch match, HookDefinition hook, HookType hookType,
+    private Runtime createRuntimeWithMockedGlue(PickleStepDefinitionMatch match, HookDefinition hook, HookType hookType,
                                                 String... runtimeArgs) {
         return createRuntimeWithMockedGlue(match, false, hook, hookType, runtimeArgs);
     }
 
     private Runtime createRuntimeWithMockedGlueWithAmbiguousMatch(String... runtimeArgs) {
-        return createRuntimeWithMockedGlue(mock(StepDefinitionMatch.class), true, mock(HookDefinition.class), HookType.After, runtimeArgs);
+        return createRuntimeWithMockedGlue(mock(PickleStepDefinitionMatch.class), true, mock(HookDefinition.class), HookType.After, runtimeArgs);
     }
 
-    private Runtime createRuntimeWithMockedGlue(StepDefinitionMatch match, boolean isAmbiguous, HookDefinition hook,
+    private Runtime createRuntimeWithMockedGlue(PickleStepDefinitionMatch match, boolean isAmbiguous, HookDefinition hook,
                                                 HookType hookType, String... runtimeArgs) {
         ResourceLoader resourceLoader = mock(ResourceLoader.class);
         ClassLoader classLoader = mock(ClassLoader.class);
@@ -552,7 +552,7 @@ public class RuntimeTest {
         return new Runtime(resourceLoader, classLoader, backends, runtimeOptions, glue);
     }
 
-    private void mockMatch(RuntimeGlue glue, StepDefinitionMatch match, boolean isAmbiguous) {
+    private void mockMatch(RuntimeGlue glue, PickleStepDefinitionMatch match, boolean isAmbiguous) {
         if (isAmbiguous) {
             Exception exception = new AmbiguousStepDefinitionsException(mock(PickleStep.class), Arrays.asList(match, match));
             doThrow(exception).when(glue).stepDefinitionMatch(anyString(), (PickleStep) any());
