+++ core/src/test/java/cucumber/runtime/StepDefinitionMatchTest.java
@@ -37,7 +37,7 @@ public class StepDefinitionMatchTest {
         PickleStep stepWithoutDocStringOrTable = mock(PickleStep.class);
         when(stepWithoutDocStringOrTable.getArgument()).thenReturn(Collections.<gherkin.pickles.Argument>emptyList());
 
-        StepDefinitionMatch stepDefinitionMatch = new StepDefinitionMatch(singletonList(createArgument(0, "5")), stepDefinition, "some.feature", stepWithoutDocStringOrTable, new LocalizedXStreams(classLoader));
+        PickleStepDefinitionMatch stepDefinitionMatch = new PickleStepDefinitionMatch(singletonList(createArgument(0, "5")), stepDefinition, "some.feature", stepWithoutDocStringOrTable, new LocalizedXStreams(classLoader));
         stepDefinitionMatch.runStep(ENGLISH, null);
         verify(stepDefinition).execute(ENGLISH, new Object[]{5});
     }
@@ -52,7 +52,7 @@ public class StepDefinitionMatchTest {
         PickleStep stepWithoutDocStringOrTable = mock(PickleStep.class);
         when(stepWithoutDocStringOrTable.getArgument()).thenReturn(Collections.<gherkin.pickles.Argument>emptyList());
 
-        StepDefinitionMatch stepDefinitionMatch = new StepDefinitionMatch(singletonList(createArgument(0, "the thing")), stepDefinition, "some.feature", stepWithoutDocStringOrTable, new LocalizedXStreams(classLoader));
+        PickleStepDefinitionMatch stepDefinitionMatch = new PickleStepDefinitionMatch(singletonList(createArgument(0, "the thing")), stepDefinition, "some.feature", stepWithoutDocStringOrTable, new LocalizedXStreams(classLoader));
         stepDefinitionMatch.runStep(ENGLISH, null);
         verify(stepDefinition).execute(ENGLISH, new Object[]{new Thing("the thing")});
     }
@@ -68,7 +68,7 @@ public class StepDefinitionMatchTest {
         PickleString docString = new PickleString(mock(PickleLocation.class), "the thing");
         when(stepWithDocString.getArgument()).thenReturn(asList((gherkin.pickles.Argument)docString));
 
-        StepDefinitionMatch stepDefinitionMatch = new StepDefinitionMatch(new ArrayList<Argument>(), stepDefinition, "some.feature", stepWithDocString, new LocalizedXStreams(classLoader));
+        PickleStepDefinitionMatch stepDefinitionMatch = new PickleStepDefinitionMatch(new ArrayList<Argument>(), stepDefinition, "some.feature", stepWithDocString, new LocalizedXStreams(classLoader));
         stepDefinitionMatch.runStep(ENGLISH, null);
         verify(stepDefinition).execute(ENGLISH, new Object[]{new Thing("the thing")});
     }
@@ -117,7 +117,7 @@ public class StepDefinitionMatchTest {
         PickleStep stepWithoutDocStringOrTable = mock(PickleStep.class);
         when(stepWithoutDocStringOrTable.getArgument()).thenReturn(Collections.<gherkin.pickles.Argument>emptyList());
 
-        StepDefinitionMatch stepDefinitionMatch = new StepDefinitionMatch(singletonList(createArgument(0, "blah")), stepDefinition, "some.feature", stepWithoutDocStringOrTable, new LocalizedXStreams(classLoader));
+        PickleStepDefinitionMatch stepDefinitionMatch = new PickleStepDefinitionMatch(singletonList(createArgument(0, "blah")), stepDefinition, "some.feature", stepWithoutDocStringOrTable, new LocalizedXStreams(classLoader));
         try {
 
             stepDefinitionMatch.runStep(ENGLISH, null);
@@ -149,7 +149,7 @@ public class StepDefinitionMatchTest {
         PickleString docString = new PickleString(mock(PickleLocation.class), "HELLO");
         when(stepWithDocString.getArgument()).thenReturn(asList((gherkin.pickles.Argument)docString));
 
-        StepDefinitionMatch stepDefinitionMatch = new StepDefinitionMatch(new ArrayList<Argument>(), stepDefinition, "some.feature", stepWithDocString, new LocalizedXStreams(classLoader));
+        PickleStepDefinitionMatch stepDefinitionMatch = new PickleStepDefinitionMatch(new ArrayList<Argument>(), stepDefinition, "some.feature", stepWithDocString, new LocalizedXStreams(classLoader));
         stepDefinitionMatch.runStep(ENGLISH, null);
         verify(stepDefinition).execute(ENGLISH, new Object[]{"HELLO"});
     }
@@ -167,7 +167,7 @@ public class StepDefinitionMatchTest {
         PickleString docString = new PickleString(mock(PickleLocation.class), "HELLO");
         when(stepWithDocString.getArgument()).thenReturn(asList((gherkin.pickles.Argument)docString));
 
-        StepDefinitionMatch stepDefinitionMatch = new StepDefinitionMatch(singletonList(createArgument(0, "5")), stepDefinition, "some.feature", stepWithDocString, new LocalizedXStreams(classLoader));
+        PickleStepDefinitionMatch stepDefinitionMatch = new PickleStepDefinitionMatch(singletonList(createArgument(0, "5")), stepDefinition, "some.feature", stepWithDocString, new LocalizedXStreams(classLoader));
         stepDefinitionMatch.runStep(ENGLISH, null);
         verify(stepDefinition).execute(ENGLISH, new Object[]{5, "HELLO"});
     }
@@ -177,7 +177,7 @@ public class StepDefinitionMatchTest {
         PickleStep step = new PickleStep("I have 4 cukes in my belly", Collections.<gherkin.pickles.Argument>emptyList(), asList(mock(PickleLocation.class)));
 
         StepDefinition stepDefinition = new StubStepDefinition(new Object(), Object.class.getMethod("toString"), "some pattern");
-        StepDefinitionMatch stepDefinitionMatch = new StepDefinitionMatch(singletonList(createArgument(7, "4")), stepDefinition, null, step, new LocalizedXStreams(getClass().getClassLoader()));
+        PickleStepDefinitionMatch stepDefinitionMatch = new PickleStepDefinitionMatch(singletonList(createArgument(7, "4")), stepDefinition, null, step, new LocalizedXStreams(getClass().getClassLoader()));
         try {
             stepDefinitionMatch.runStep(ENGLISH, null);
             fail();
@@ -197,7 +197,7 @@ public class StepDefinitionMatchTest {
         PickleStep step = new PickleStep("I have 4 cukes in my belly", asList((gherkin.pickles.Argument)mock(PickleTable.class)), asList(mock(PickleLocation.class)));
 
         StepDefinition stepDefinition = new StubStepDefinition(new Object(), WithTwoParams.class.getMethod("withTwoParams", Integer.TYPE, Short.TYPE, List.class), "some pattern");
-        StepDefinitionMatch stepDefinitionMatch = new StepDefinitionMatch(singletonList(createArgument(7, "4")), stepDefinition, null, step, new LocalizedXStreams(getClass().getClassLoader()));
+        PickleStepDefinitionMatch stepDefinitionMatch = new PickleStepDefinitionMatch(singletonList(createArgument(7, "4")), stepDefinition, null, step, new LocalizedXStreams(getClass().getClassLoader()));
         try {
             stepDefinitionMatch.runStep(ENGLISH, null);
             fail();
